requirements.txt
uvicorn
numpy
tensorflow-cpu
fastapi
gunicorn==19.7.1
Jinja2==2.8.1
